+++ src/main/java/roomescape/common/config/ReservationConfig.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.common.argumentResolver.LoginArgumentResolver;
-import roomescape.common.interceptor.AdminInterceptor;
+import roomescape.member.argumentResolver.LoginArgumentResolver;
+import roomescape.admin.interceptor.AdminInterceptor;
 import roomescape.common.util.JwtTokenContainer;
 import roomescape.common.util.TokenCookieManager;
 import roomescape.member.service.LoginService;
